+++ Asteroids/Assets.h
@@ -14,12 +14,12 @@ public:
 	sf::Font& getFont();
 	sf::Texture& getPlayerTexture();
 	sf::Texture& getAsteroidsTexture();
-	std::unordered_map<std::string, int>* readGameParameters(const std::string& filename);
-	std::unordered_map<std::string, int>& getGameParams();
+	std::unordered_map<std::string, float>* readGameParameters(const std::string& filename);
+	std::unordered_map<std::string, float>& getGameParams();
 	
 private:
 	sf::Texture* PlayerTexture = new sf::Texture();
 	sf::Texture* AsteroidTexture = new sf::Texture();
 	sf::Font* font = new sf::Font();
-	std::unordered_map<std::string, int>* params;
+	std::unordered_map<std::string, float>* params;
 };
